third_party/WebKit/LayoutTests/fast/dom/Window/alert-undefined-expected.txt

ALERT: undefined
ALERT: null
CONFIRM: undefined
CONFIRM: null
PROMPT: undefined, default text: 
PROMPT: null, default text: 
ALERT: 
CONFIRM: 
PROMPT: , default text: 
